+++ src/test/java/com/vaadin/ui/paper/dialog/GeneratedPaperDialog.java
